Refactor content relations API. #206
