+++ src/API/ContentRelations.php
@@ -1,42 +1,16 @@
 <?php # -*- coding: utf-8 -*-
+
+namespace Inpsyde\MultilingualPress\API;
+
+// TODO: For now, this is just a copy of the old API. Functionally refactor this as soon as the structural one is done.
+
 /**
- * Relationships between content blocks (posts, terms, whatever).
+ * Interface for all content relations API implementations.
  *
- * @version 2015.07.30
- * @author  Inpsyde GmbH, toscho, tf
- * @license GPL
+ * @package Inpsyde\MultilingualPress\API
+ * @since   3.0.0
  */
-interface Mlp_Content_Relations_Interface {
-
-	/**
-	 * Set a relation according to the given parameters.
-	 *
-	 * @param int    $source_site_id    Source blog ID.
-	 * @param int    $target_site_id    Target blog ID.
-	 * @param int    $source_content_id Source post ID or term taxonomy ID.
-	 * @param int    $target_content_id Target post ID or term taxonomy ID.
-	 * @param string $type              Content type.
-	 *
-	 * @return bool
-	 */
-	public function set_relation(
-		$source_site_id,
-		$target_site_id,
-		$source_content_id,
-		$target_content_id,
-		$type = 'post'
-	);
-
-	/**
-	 * Return an array with site IDs as keys and content IDs as values.
-	 *
-	 * @param int    $source_site_id    Source blog ID.
-	 * @param int    $source_content_id Source post ID or term taxonomy ID.
-	 * @param string $type              Content type.
-	 *
-	 * @return array
-	 */
-	public function get_relations( $source_site_id, $source_content_id, $type = 'post' );
+interface ContentRelations {
 
 	/**
 	 * Delete a relation according to the given parameters.
@@ -58,21 +32,19 @@ interface Mlp_Content_Relations_Interface {
 	);
 
 	/**
-	 * Return the existing (or new) translation IDs according to the given parameters.
+	 * Return the term taxonomy ID of the given target site for the given source term.
 	 *
 	 * @param int    $source_site_id    Source blog ID.
 	 * @param int    $target_site_id    Target blog ID.
 	 * @param int    $source_content_id Source post ID or term taxonomy ID.
-	 * @param int    $target_content_id Target post ID or term taxonomy ID.
 	 * @param string $type              Content type.
 	 *
-	 * @return array
+	 * @return int
 	 */
-	public function get_translation_ids(
+	public function get_element_for_site(
 		$source_site_id,
 		$target_site_id,
 		$source_content_id,
-		$target_content_id,
 		$type
 	);
 
@@ -96,20 +68,51 @@ interface Mlp_Content_Relations_Interface {
 	);
 
 	/**
-	 * Return the term taxonomy ID of the given target site for the given source term.
+	 * Return an array with site IDs as keys and content IDs as values.
+	 *
+	 * @param int    $source_site_id    Source blog ID.
+	 * @param int    $source_content_id Source post ID or term taxonomy ID.
+	 * @param string $type              Content type.
+	 *
+	 * @return array
+	 */
+	public function get_relations( $source_site_id, $source_content_id, $type = 'post' );
+
+	/**
+	 * Return the existing (or new) translation IDs according to the given parameters.
 	 *
 	 * @param int    $source_site_id    Source blog ID.
 	 * @param int    $target_site_id    Target blog ID.
 	 * @param int    $source_content_id Source post ID or term taxonomy ID.
+	 * @param int    $target_content_id Target post ID or term taxonomy ID.
 	 * @param string $type              Content type.
 	 *
-	 * @return int
+	 * @return array
 	 */
-	public function get_element_for_site(
+	public function get_translation_ids(
 		$source_site_id,
 		$target_site_id,
 		$source_content_id,
+		$target_content_id,
 		$type
 	);
 
+	/**
+	 * Set a relation according to the given parameters.
+	 *
+	 * @param int    $source_site_id    Source blog ID.
+	 * @param int    $target_site_id    Target blog ID.
+	 * @param int    $source_content_id Source post ID or term taxonomy ID.
+	 * @param int    $target_content_id Target post ID or term taxonomy ID.
+	 * @param string $type              Content type.
+	 *
+	 * @return bool
+	 */
+	public function set_relation(
+		$source_site_id,
+		$target_site_id,
+		$source_content_id,
+		$target_content_id,
+		$type = 'post'
+	);
 }
